Create sroc and presroc `ViewInvoiceService`
We copy the existing `ViewInvoiceService` into separate sroc and presroc versions, then update `ViewInvoiceService` to call the presroc version for now.
