+++ app/services/invoices/view_invoice.service.js
@@ -4,8 +4,7 @@
  * @module ViewInvoiceService
  */
 
-const { ViewInvoicePresenter } = require('../../presenters')
-const FetchAndValidateInvoiceService = require('./fetch_and_validate_invoice.service')
+const ViewInvoicePresrocService = require('./presroc/view_invoice_presroc.service')
 
 class ViewInvoiceService {
   /**
@@ -18,78 +17,7 @@ class ViewInvoiceService {
    * @returns {Object} The requested invoice data
    */
   static async go (billRunId, invoiceId) {
-    let invoice = await FetchAndValidateInvoiceService.go(billRunId, invoiceId)
-
-    invoice = await this._invoiceResponseData(invoice)
-
-    return this._response(invoice)
-  }
-
-  static async _invoiceResponseData (invoice) {
-    const responseData = await invoice.$query()
-      .select(
-        'id',
-        'billRunId',
-        'customerReference',
-        'financialYear',
-        'deminimisInvoice',
-        'zeroValueInvoice',
-        'minimumChargeInvoice',
-        'transactionReference',
-        'debitLineValue',
-        'creditLineValue',
-        'rebilledType',
-        'rebilledInvoiceId'
-      )
-      .withGraphFetched('licences.transactions')
-      .modifyGraph('licences', builder => {
-        builder.select(
-          'id',
-          'licenceNumber',
-          'debitLineValue',
-          'creditLineValue'
-        )
-      })
-      .modifyGraph('licences.transactions', builder => {
-        builder.select(
-          'id',
-          'clientId',
-          'chargeValue',
-          'chargeCredit',
-          'subjectToMinimumCharge',
-          'minimumChargeAdjustment',
-          'lineDescription',
-          'chargePeriodStart',
-          'chargePeriodEnd',
-          'regimeValue17',
-          'rebilledTransactionId',
-          'chargeCalculation'
-        )
-      })
-
-    return {
-      ...responseData,
-      netTotal: responseData.$netTotal(),
-      licences: this._addNetTotalToLicence(responseData.licences)
-    }
-  }
-
-  static _response (invoice) {
-    const presenter = new ViewInvoicePresenter(invoice)
-
-    return presenter.go()
-  }
-
-  /**
-   * Take an array of licences and adds $netTotal() to each one then returns the resulting array
-   */
-  static _addNetTotalToLicence (licences) {
-    return licences.map(licence => {
-      return {
-        ...licence,
-        netTotal: licence.$netTotal()
-      }
-    })
+    return ViewInvoicePresrocService.go(billRunId, invoiceId)
   }
 }
